fix: satisfy backend-size limits for restart-resume mergeability
Split oversized factorysessionexecution files and resume fixture helpers so
Build, Lint, and API backend-size checks pass on the PR head.
